train_model/requirements.txt
craft-text-detector
opencv-python
python_dotenv
ultralytics
vietocr